Move MatterCommissionerControlPluginServerInitCallback from application to SDK
