+++ examples/fabric-bridge-app/fabric-bridge-common/src/ZCLCallbacks.cpp
@@ -20,7 +20,6 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Attributes.h>
 #include <app-common/zap-generated/ids/Clusters.h>
-#include <app/clusters/commissioner-control-server/commissioner-control-server.h>
 #include <lib/support/ZclString.h>
 
 using namespace ::chip;
@@ -90,14 +89,4 @@ Protocols::InteractionModel::Status emberAfExternalAttributeWriteCallback(Endpoi
     }
 
     return ret;
-}
-
-void MatterCommissionerControlPluginServerInitCallback()
-{
-    ChipLogProgress(Zcl, "MatterCommissionerControlPluginServerInitCallback");
-
-    BitMask<CommissionerControl::SupportedDeviceCategoryBitmap> supportedDeviceCategories;
-    supportedDeviceCategories.SetField(CommissionerControl::SupportedDeviceCategoryBitmap::kFabricSynchronization, 1);
-    CommissionerControl::CommissionerControlServer::Instance().SetSupportedDeviceCategoriesValue(kRootEndpointId,
-                                                                                                 supportedDeviceCategories);
 }
