+++ src/app/clusters/commissioner-control-server/commissioner-control-server.h
@@ -82,7 +82,7 @@ public:
      * @brief
      *   Called after the server return SUCCESS to a correctly formatted RequestCommissioningApproval command.
      */
-    CHIP_ERROR EmitCommissioningRequestResultEvent(const Events::CommissioningRequestResult::Type & result);
+    CHIP_ERROR GenerateCommissioningRequestResultEvent(const Events::CommissioningRequestResult::Type & result);
 
 private:
     CommissionerControlServer()  = default;
